Fix app.go panic being ignored silently
The way output hijacking was set up for app.go, a panic inside app.go
would be silently ignored because the tail.Tee in the wrapper would stop
too early--largely because app.go would send the signal that it was done
before it had written its panic to the output stream.

To address this, the output hijacking and send signal have to happen a
lot closer to the entry point in main. This messes up main.go a tad bit
in terms of cleaniliness but it ensures that the output makes it
through.