Add integration test (#25)
Add an integration test that builds and runs tmux-fastcopy with
different versions of tmux. Only 3.2a for now because we use the `-e`
flag on `new-session` which was added in 3.2.

The test itself is fairly messy because it needs to simulate a VT100
terminal to make assertions about the output.