Handle line wrapping
When a line is longer than the width of the pane,
tmux wraps it like so:

    |foob|
    |ar  |

`tmux capture-pane` reports these as two separate lines by default.

    $ tmux capture-pane
    foob
    ar

This means that really long strings that would otherwise be matched by
tmux-fastcopy get split and possibly ignored.

We can address this by using the `-J` flag of `tmux capture-pane`,
which provides a more accurate representation of the actual text
on-screen.

    $ tmux capture-pane -J
    foobar

Combining that with some basic record-keeping of the screen width, we
can show the user the wrapped text, and still copy the actual unwrapped
text.

Testing this required making the integration test slightly more robust.