integration test: Count for coverage (#68)
The integration test works by spawning a tmux session in a fake
terminal, which then spawns copies of tmux-fastcopy. To get control of
the executable spawned by tmux, we make it spawn the test binary and
hijack it in the TestMain. This provides access to a
coverage-instrumented copy of tmux-fastcopy's main.

To extract coverage out of this, use `"unsafe"` and `go:linkname` to get
access to two private components of the testing package: the `cover`
variable that holds the coverage data for the current binary, and the
`coverReport` function which reports the coverage for the current
binary.

With these two pieces, coverage extraction works like so:

- tmux spawns a coverage-instrumented tmux-fastcopy, which uses
  coverReport to place its coverage data into a pre-determined directory
- the test, after running, reads all of these coverage files and
  incoprorates data from them back into the coverage data for the
  current binary

This is all inspired by how rogpeppe's [go-internal/testscript][1]
operates, but rather than recreate the full testing machinery,
we use specific parts of it with `go:linkname`.

[1]: https://github.com/rogpeppe/go-internal/blob/3461ca1f2345421c6d6d05407a3ac0381bbd5c42/testscript/cover.go