ui/run: Fix data race between shutdown and drawing
There's a data race when an event to exit is received.
When handleEvent receives the signal to exit,
it'll tell tcell to start cleaning up,
which happens in a separate goroutine.
At the same time, we'll call screen.Clear and draw on it.

Fix by drawing only if a non-exit event was received.