test: Fix integration test
The integration test has been partially broken for a while.
It's not clear why, but it stopped "seeing" the red text.
The vt100 library's view of the world includes
the white text and the green text next to the hints,
but not the red hints.

To fix this, replace the complex vt100 simulation with a much simpler
buffer matching.
This does require us to know the exact codes of the hints that we're
looking for, but it's not too difficult.