huffman: optimize label assignment slightly (#209)
The huffman algorithm works by building a tree
where each node has up to $base children (for a $base alphabet).
High frequency nodes are near the root of the tree,
and low frequency nodes are near the bottom.
Path from root to leaf nodes is the label for that item.

Previously, when building the tree,
we would cluster the items with the lowest frequency
into a node of size $base.
This would result in cases where the topmost nodes (highest frequency)
have fewer than $base children.

So when we assign labels, a few items end up in a lower level
than they could given their frequency and room in the topmost nodes.
So e.g. if there are 4 nodes (A-D), and base is 3. we create the tree:

         *
        /|
      0/ |1
      D  *
        /|\
      0/ |1\2
      A  B  C

This results in the labels:

    D: 0
    A: 10
    B: 11
    C: 12

This PR changes the algorithm slightly so that topmost node
always has $base children,
and the first cluster we create (the lowest frequency nodes)
gets < $base children if that's the case we're in.

As a result, we'll build the tree:

          *
         /|\
       0/ |1\2
        D A  *
            /|
          0/ |1
          B  C

This results in the labels:

    D: 0
    A: 1
    B: 20
    C: 21

More nodes get single-letter labels than before.